Move helper modules into their own subpackage
