+++ src/roles/gunner.py
@@ -12,7 +12,7 @@ from src.containers import UserList, UserSet, UserDict, DefaultUserDict
 from src.messages import messages
 from src.events import Event
 
-from src.roles._gunner_helper import setup_variables
+from src.roles.helper.gunners import setup_variables
 
 GUNNERS = setup_variables("gunner")
