+++ src/utilities.py
@@ -48,15 +48,15 @@ def reply(cli, nick, chan, msg, private=False, prefix_nick=False):
         else:
             cli.msg(chan, msg)
 
-# FIXME: Deprecated in favor of _wolf_helper method
+# FIXME: Deprecated in favor of helper.wolves method
 def in_wolflist(nick, who):
-    from src.roles._wolf_helper import is_known_wolf_ally
+    from src.roles.helper.wolves import is_known_wolf_ally
     from src import users
     return is_known_wolf_ally(var, users._get(nick), users._get(who))
 
-# FIXME: Deprecated in favor of _wolf_helper method
+# FIXME: Deprecated in favor of helper.wolves method
 def relay_wolfchat_command(cli, nick, message, roles, is_wolf_command=False, is_kill_command=False):
-    from src.roles._wolf_helper import send_wolfchat_message
+    from src.roles.helper.wolves import send_wolfchat_message
     from src import users
     role = "wolf" if is_wolf_command else None
     command = "kill" if is_kill_command else None
